Added hearts the restore player health!
Bug: Hearts still show up after being used, although they only heal
once.

Modified PlayerHealth to heal when a heart is picked up.
Added method addHeart() to player to be able to add hearts to a level.
Added a heart on Level1 to test this functionality.
Added HeartRender which draws hearts.
Added HeartPhysics so hearts fall to the floor.
Added heart.png to img so hearts have a sprite.

Change-Id: I67243313bdbda9ad7e014c380ab7fc4877dffc7e